requirements.txt
pytest
requests
allure-pytest==2.8.6
allure-python-commons==2.8.6
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
configparser==4.0.2
crayons==0.3.0
future==0.18.2
git-pylint-commit-hook==2.5.1
idna==2.8
importlib-metadata==1.3.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.0.2
packaging==19.2
pluggy==0.13.1
py==1.8.0
pycodestyle==2.5.0
pylint==2.4.4
pyodbc==4.0.28
pyparsing==2.4.5
pytest==5.3.2
requests==2.22.0
selenium==3.141.0
six==1.13.0
typed-ast==1.4.1
urllib3==1.25.7
wcwidth==0.1.7
webdriver-manager==2.3.0
wrapt==1.11.2
zipp==0.6.0